Move VideoPage under Gatsby component and add Layout
