+++ src/tmp/index.ts
@@ -1,8 +1,6 @@
 import Dashboard from './Dashboard/Dashboard'
-import VideoPage from './VideoPlayer/VideoPage'
 
 export {
-    Dashboard,
-    VideoPage,
+    Dashboard
 }
